requirements.txt
streamlit==1.28.0
google-generativeai==0.3.0
firebase-admin==6.2.0
Pillow==10.1.0
python-dotenv==1.0.0
moviepy==1.0.3
decorator==4.4.2
opencv-python-headless==4.9.0.80
numpy==1.26.4
